refactor: move status publishing into its own file, rename resolved/unresolved to assigned/published
Superwall.swift kept growing a small state machine named in jargon. The
pipeline now lives in SubscriptionStatusPublishing.swift with one doc
comment explaining the two values: the assigned status (what a writer
handed the SDK, persisted) and the published status (assigned + granted
entitlements + test-mode override, what subscriptionStatus holds).

Five functions collapse to three — publishSubscriptionStatus,
mergedSubscriptionStatus, and the automatic-path customer info refresh —
with the developer-assignment warning inlined into publish. The class
keeps only the @Published property with its observers and the stored
state the pipeline needs. setSubscriptionStatus(base:) becomes
setSubscriptionStatus(assigned:). No behaviour change; tests renamed to
match.
